-Fix cast warnings -Fix signed/unsigned comparison warnings
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@7259 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/dls.cpp
@@ -221,7 +221,7 @@ MDLS *DLS_get_access(void)
 	if (!is_initialized) {
 		is_initialized = true;
 		TempDirectoryList dir_list;
-		for (int i = 0; i < dir_list.Count(); i++) {
+		for (size_t i = 0; i < dir_list.Count(); i++) {
 			TempDirectoryList::Item item = dir_list[i];
 			DLS_add_dir(item.size, item.dir.c_str());
 		}
